requirements/local.txt
celery==3.1.23
Django==1.8
django-celery==3.1.17
django-cors-headers==1.1.0
django-debug-panel==0.8.3
django-debug-toolbar==1.4
django-rest-framework-mongoengine==3.3.0
django-haystack==2.4.1
django-smtp-ssl==1.0
django-user-agents==0.3.0
djangorestframework==3.3.3
gunicorn==19.4.5
ipdb==0.9.3
mongoengine==0.10.6
pathlib2==2.1.0
Pillow==3.2.0
# psycopg2==2.6.1
ptyprocess==0.5.1
pymongo==3.2.2
pytz==2016.4
redis==2.10.5
requests==2.10.0
supervisor==3.3.0
ua-parser==0.7.1
user-agents==1.0.1